requirements.txt
# Core dependencies
streamlit>=1.22.0
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# PDF and document processing
pypdf>=3.15.1
pdfminer.six>=20221105
pandas>=2.0.0
numpy>=1.24.0

# Vector database
faiss-cpu>=1.7.4
chromadb>=0.4.13

# Utilities
tqdm>=4.66.1
tiktoken>=0.5.1
tenacity>=8.2.3 